+++ tests/codeception/unit/entry/CalendarEntryFormSequenceTest.php
@@ -29,14 +29,17 @@ class CalendarEntryFormSequenceTest extends CalendarUnitTest
                 'all_day' => '0',
                 'title' => 'Test title',
                 'description' => 'TestDescription',
-                'participation_mode' => 2
+                'participation_mode' => 2,
+                'recurring' => 1,
             ],
             'CalendarEntryForm' => [
                 'is_public' => '1',
                 'start_date' => '6/27/17',
                 'start_time' => '12:00 PM',
                 'end_date' => '6/28/17',
-                'end_time' => '01:00 PM'
+                'end_time' => '01:00 PM',
+                'reminder' => 1,
+                'recurring' => 1,
             ]
         ]));
 
@@ -53,7 +56,9 @@ class CalendarEntryFormSequenceTest extends CalendarUnitTest
             ],
             'CalendarEntryForm' => [
                 'start_date' => '7/16/17',
-                'end_date' => '8/16/17'
+                'end_date' => '8/16/17',
+                'reminder' => 1,
+                'recurring' => 1,
             ]
         ]);
 
@@ -70,7 +75,9 @@ class CalendarEntryFormSequenceTest extends CalendarUnitTest
                 'all_day' => '1'
             ],
             'CalendarEntryForm' => [
-                'end_date' => '8/16/17'
+                'end_date' => '8/16/17',
+                'reminder' => 1,
+                'recurring' => 1,
             ]
         ]);
 
@@ -84,7 +91,10 @@ class CalendarEntryFormSequenceTest extends CalendarUnitTest
         $calendarForm = new CalendarEntryForm(['entry' => $this->form->entry]);
         $calendarForm->load([
             'CalendarEntry' => [],
-            'CalendarEntryForm' => [],
+            'CalendarEntryForm' => [
+                'reminder' => 1,
+                'recurring' => 1,
+            ],
             'RecurrenceFormModel' => [
                 'frequency' => Frequency::DAILY,
                 'interval' => 1
@@ -102,7 +112,9 @@ class CalendarEntryFormSequenceTest extends CalendarUnitTest
         $calendarForm->load([
             'CalendarEntry' => [],
             'CalendarEntryForm' => [
-                'is_public' => '0'
+                'is_public' => '0',
+                'reminder' => 1,
+                'recurring' => 1,
             ],
         ]);