+++ views/common/_settings_participation.php
@@ -1,9 +1,10 @@
 <?php
 
-use humhub\components\View;
-use humhub\modules\calendar\models\forms\CalendarEntryForm;
+use humhub\modules\calendar\assets\ParticipationFormAssets;
+use humhub\modules\calendar\models\forms\CalendarEntryParticipationForm;
 use humhub\modules\calendar\models\participation\ParticipationSettings;
 use humhub\widgets\Button;
+use yii\web\View;
 
 /* @var $this View */
 /* @var $participationSettings ParticipationSettings */
@@ -12,9 +13,10 @@ $helpBlock = $participationSettings->isGlobal()
     ? Yii::t('CalendarModule.config', 'Here you can configure default settings for new calendar events. These settings can be overwritten on space/profile level.')
     : Yii::t('CalendarModule.config', 'Here you can configure default settings for new calendar events.') ;
 
+ParticipationFormAssets::register($this);
 ?>
 
-<div class="panel-body" data-ui-widget="calendar.Form">
+<div class="panel-body" data-ui-widget="calendar.participation.Form">
     <h4>
         <?= Yii::t('CalendarModule.config', 'Default participation settings'); ?>
         <?php if ($participationSettings->showResetButton()) : ?>
@@ -27,7 +29,7 @@ $helpBlock = $participationSettings->isGlobal()
         <?= $helpBlock ?>
     </div>
 
-    <?= $form->field($participationSettings, 'participation_mode')->dropDownList(CalendarEntryForm::getParticipationModeItems(), ['data-action-change' => 'changeParticipationMode']) ?>
+    <?= $form->field($participationSettings, 'participation_mode')->dropDownList(CalendarEntryParticipationForm::getModeItems(), ['data-action-change' => 'changeParticipationMode']) ?>
     <div class="participationOnly" style="<?= $participationSettings->isParticipationAllowed() ? '' : 'display:none' ?>">
         <?= $form->field($participationSettings, 'allow_decline')->checkbox() ?>
         <?= $form->field($participationSettings, 'allow_maybe')->checkbox() ?>