+++ views/entry/edit-basic.php
@@ -2,11 +2,13 @@
 
 use humhub\modules\calendar\models\CalendarEntryType;
 use humhub\modules\calendar\models\forms\CalendarEntryForm;
+use humhub\modules\calendar\Module;
 use humhub\modules\content\components\ContentContainerActiveRecord;
 use humhub\modules\content\widgets\richtext\RichTextField;
 use humhub\modules\topic\widgets\TopicPicker;
 use humhub\modules\ui\form\widgets\ColorPicker;
 use humhub\modules\content\widgets\ContentTagDropDown;
+use humhub\modules\ui\form\widgets\ContentVisibilitySelect;
 use humhub\modules\ui\form\widgets\TimePicker;
 use humhub\widgets\ActiveForm;
 use humhub\widgets\TimeZoneDropdownAddition;
@@ -65,11 +67,26 @@ use yii\jui\DatePicker;
 
     <?php Yii::$app->i18n->autosetLocale(); ?>
 
-    <?= $form->field($calendarEntryForm->entry, 'all_day')->checkbox(['data-action-change' => 'toggleDateTime']) ?>
-
-    <?php if($calendarEntryForm->canCreatePublicEntry()) :?>
-        <?= $form->field($calendarEntryForm, 'is_public')->checkbox() ?>
-    <?php endif; ?>
+    <div class="row">
+        <div class="col-sm-4">
+            <?= $form->field($calendarEntryForm->entry, 'all_day')->checkbox(['data-action-change' => 'toggleDateTime']) ?>
+        </div>
+        <?php if (Module::isRecurrenceActive()) : ?>
+        <div class="col-sm-4">
+            <?= $form->field($calendarEntryForm, 'recurring')->checkbox(['data-action-change' => 'toggleRecurring']) ?>
+        </div>
+        <?php endif; ?>
+        <?php if ($calendarEntryForm->isFutureEvent()) : ?>
+        <div class="col-sm-4">
+            <?= $form->field($calendarEntryForm, 'reminder')->checkbox(['data-action-change' => 'toggleReminder']) ?>
+        </div>
+        <?php endif; ?>
+        <?php if ($calendarEntryForm->canCreatePublicEntry()) :?>
+        <div class="col-sm-12">
+            <?= $form->field($calendarEntryForm, 'is_public')->widget(ContentVisibilitySelect::class, ['contentOwner' => 'entry']) ?>
+        </div>
+        <?php endif; ?>
+    </div>
 
     <?= $form->field($calendarEntryForm->entry, 'description')->widget(RichTextField::class, ['placeholder' => Yii::t('CalendarModule.base', 'Description'), 'pluginOptions' => ['maxHeight' => '300px']])->label(false) ?>
 
@@ -82,4 +99,8 @@ use yii\jui\DatePicker;
     ])->label(false) ?>
 
     <?= $form->field($calendarEntryForm, 'topics')->widget(TopicPicker::class, ['contentContainer' => $contentContainer]); ?>
+
+    <?php if (!$calendarEntryForm->entry->isNewRecord) : ?>
+        <?= $form->field($calendarEntryForm, 'sendUpdateNotification')->checkbox() ?>
+    <?php endif; ?>
 </div>