+++ views/entry/edit.php
@@ -1,11 +1,7 @@
 <?php
-
-
 use humhub\modules\calendar\assets\CalendarBaseAssets;
 use humhub\modules\calendar\models\forms\CalendarEntryForm;
 use humhub\modules\calendar\helpers\RecurrenceHelper;
-use humhub\modules\calendar\Module;
-use humhub\widgets\Button;
 use humhub\widgets\ModalButton;
 use humhub\widgets\Tabs;
 use humhub\modules\ui\form\widgets\ActiveForm;
@@ -14,12 +10,17 @@ use humhub\widgets\ModalDialog;
 /* @var $this \humhub\modules\ui\view\components\View */
 /* @var $calendarEntryForm CalendarEntryForm */
 /* @var $contentContainer \humhub\modules\content\components\ContentContainerActiveRecord */
+/* @var $editUrl string */
 
 CalendarBaseAssets::register($this);
 
-$header = ($calendarEntryForm->entry->isNewRecord)
-    ? Yii::t('CalendarModule.views_entry_edit', '<strong>Create</strong> event')
-    : Yii::t('CalendarModule.views_entry_edit', '<strong>Edit</strong> event');
+if ($calendarEntryForm->entry->isNewRecord) {
+    $header = Yii::t('CalendarModule.views_entry_edit', '<strong>Create</strong> Event');
+    $saveButtonText = Yii::t('CalendarModule.views_entry_edit', 'Next');
+} else {
+    $header = Yii::t('CalendarModule.views_entry_edit', '<strong>Edit</strong> Event');
+    $saveButtonText = null;
+}
 
 if(RecurrenceHelper::isRecurrent($calendarEntryForm->entry)) {
     $header = Yii::t('CalendarModule.views_entry_edit', '<strong>Edit</strong> recurring event');
@@ -30,7 +31,7 @@ $calendarEntryForm->entry->color = empty($calendarEntryForm->entry->color) ? $th
 ?>
 
 
-<?php ModalDialog::begin(['header' => $header, 'closable' => false]) ?>
+<?php ModalDialog::begin(['header' => $header, 'size' => 'large', 'closable' => false]) ?>
     <?php $form = ActiveForm::begin(['enableClientValidation' => false]); ?>
 
         <div id="calendar-entry-form" data-ui-widget="calendar.Form" data-ui-init data-is-recurrent="<?= RecurrenceHelper::isRecurrent($calendarEntryForm->entry)?>">
@@ -42,20 +43,32 @@ $calendarEntryForm->entry->color = empty($calendarEntryForm->entry->color) ? $th
                     'viewPath' => '@calendar/views/entry',
                     'params' => ['form' => $form, 'calendarEntryForm' => $calendarEntryForm, 'contentContainer' => $contentContainer],
                     'items' => [
-                        ['label' => Yii::t('CalendarModule.views_entry_edit', 'Basic'),'view' => 'edit-basic', 'linkOptions' => ['class' => 'tab-basic']],
-                        ['label' => Yii::t('CalendarModule.views_entry_edit', 'Participation'),'view' => 'edit-participation', 'linkOptions' => ['class' => 'tab-participation']],
-                        ['label' => Yii::t('CalendarModule.views_entry_edit', 'Reminder'),'view' => 'edit-reminder', 'linkOptions' => ['class' => 'tab-reminder'], 'visible' =>  ($calendarEntryForm->showReminderTab() )],
-                        ['label' => Yii::t('CalendarModule.views_entry_edit', 'Recurrence'),'view' => 'edit-recurrence', 'linkOptions' => ['class' => 'tab-recurrence'], 'visible' => Module::isRecurrenceActive()],
-                        ['label' => Yii::t('CalendarModule.views_entry_edit', 'Files'), 'view' => 'edit-files', 'linkOptions' => ['class' => 'tab-files']]
-                     ]
+                        [
+                            'label' => Yii::t('CalendarModule.views_entry_edit', 'General'),
+                            'view' => 'edit-basic',
+                            'linkOptions' => ['class' => 'tab-basic'],
+                        ],
+                        [
+                            'label' => Yii::t('CalendarModule.views_entry_edit', 'Reminder'),
+                            'view' => 'edit-reminder',
+                            'linkOptions' => ['class' => 'tab-reminder'],
+                            'headerOptions' => $calendarEntryForm->showReminderTab() ? [] : ['style' => 'display:none'],
+                        ],
+                        [
+                            'label' => Yii::t('CalendarModule.views_entry_edit', 'Recurrence'),
+                            'view' => 'edit-recurrence',
+                            'linkOptions' => ['class' => 'tab-recurrence'],
+                            'headerOptions' => $calendarEntryForm->showRecurrenceTab() ? [] : ['style' => 'display:none'],
+                        ],
+                    ]
                 ]); ?>
             </div>
 
             <hr>
 
             <div class="modal-footer">
-                <?= ModalButton::submitModal($editUrl); ?>
                 <?= ModalButton::cancel(); ?>
+                <?= ModalButton::submitModal($editUrl, $saveButtonText); ?>
             </div>
         </div>
     <?php ActiveForm::end(); ?>