+++ widgets/EntryParticipants.php
@@ -24,19 +24,21 @@ class EntryParticipants extends Widget
 
     public function run()
     {
-        if($this->calendarEntry->closed) {
-            return;
+        if ($this->calendarEntry->closed) {
+            return '';
         }
 
         $countAttending = $this->getParticipantStateCount(CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED, true);
         $countMaybe = $this->getParticipantStateCount(CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE, $this->calendarEntry->allow_maybe);
         $countDeclined = $this->getParticipantStateCount(CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED, $this->calendarEntry->allow_decline);
+        $countInvited = $this->getParticipantStateCount(CalendarEntryParticipant::PARTICIPATION_STATE_INVITED, true);
 
         return $this->render('participants', [
             'calendarEntry' => $this->calendarEntry,
             'countAttending' => $countAttending,
             'countMaybe' => $countMaybe,
             'countDeclined' => $countDeclined,
+            'countInvited' => $countInvited,
         ]);
     }
 
@@ -60,14 +62,15 @@ class EntryParticipants extends Widget
 
         $participantSate = $calendarEntry->getParticipationStatus(Yii::$app->user->identity);
 
-        $button = Button::info($label)->sm()
-            ->icon($participantSate === $state ? 'fa-check' : null);
+        $button = Button::info($label)
+            ->icon($participantSate === $state ? 'fa-check-circle' : null);
         if ($calendarEntry->isPast()) {
             $button->cssClass('disabled')->loader(false);
         } else {
-            $button->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state));
+            $button->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state))
+                ->cssClass($participantSate === $state ? '' : 'active');
         }
 
         return $button;
-    }
+   }
 }